src/app/trips-manager/requirements.txt
flask
pymongo
kafka-python
python-dotenv
requests
numpy